Ajout du calendrier & api-rest, Update discussions & listesDiscussions

+++ utilisateur/listeDiscussions.php
@@ -7,7 +7,6 @@ $objetUtilisateur = new Utilisateur($idEmploye);
 $objetMessage = new Message();
 $service = new Service();
 $discussions = $objetDiscussion->recupererDiscussions();
-$utilisateur = 0;
 ?>
 
 <div class="fleche_retour mb-2 ml-4">
@@ -40,6 +39,14 @@ if (!empty($_GET["success"]) && $_GET["success"] == "discussion") {
                 case "fonctionDiscussion": ?>
                     <?php echo "Une erreur s'est produite lors de la création de la discussion"; ?>
                     <?php break; ?>
+                    <?php
+                case "suppression": ?>
+                    <?php echo "Une erreur s'est produite lors de la suppression de la discussion"; ?>
+                    <?php break; ?>
+                    <?php
+                case "suppressionMessage": ?>
+                    <?php echo "Une erreur s'est produite lors de la suppression des messages de la discussion"; ?>
+                    <?php break; ?>
                 <?php
             }
             ?>
@@ -65,7 +72,7 @@ if (!empty($_GET["success"]) && $_GET["success"] == "discussion") {
                     </button>
                     <div class="navbar-collapse collapse" id="navbarSupportedContent11">
                     <!-- ------------------------- SELECT DESTINATAIRE ------------------------- -->
-                        <form method="post" action="../traitements/ajoutDiscussion.php?id=<?= $utilisateur; ?>">
+                        <form method="post" action="../traitements/ajoutDiscussion.php">
                             <div class="form-group">
                             <label for="idDestinataire">Envoi d'un message à : </label>
                             <select name="idDestinataire" id="idDestinataire" class="form-control">
@@ -114,7 +121,7 @@ if (!empty($_GET["success"]) && $_GET["success"] == "discussion") {
 /* ------------------------- AFFICHAGE DES DISCUSSIONS EN COURS ------------------------- */
 foreach($discussions as $discussion)
 {
-    /* Si l'utilisateur de la session a initié la discussion */
+    /* ------------ Si l'utilisateur de la session a initié la discussion ------------ */
     if($discussion["idEnvoyeur"] == $idEmploye)
     {
         $idContact = $discussion["idDestinataire"];
@@ -144,15 +151,17 @@ foreach($discussions as $discussion)
                     <div class="text-muted small">Dernière activité : <?=$service->dateFr($dernierMessage["max_date"]);?></div>
                 </div>
                 <!-- Top right -->
-                <a href="../traitements/supprimerDiscussion.php" class="icone_poubelle">
+                <a href="../traitements/supprimerDiscussion.php?idDiscussion=<?=$discussion["idDiscussion"];?>" class="icone_poubelle" onclick="return confirm('Êtes-vous sûr de vouloir supprimer la discussion ? Tous les messages seront également supprimés')">
                 <i class="far fa-trash-alt"></i>
                 </a>
             </div>
         </div>
+        <!--
         <a href="discussion.php?id=<?=$discussion["idDiscussion"];?>" class="card_discussion">
-        <div class="card-body px-2 py-1">
+                    -->
+        <div class="card-body px-2 py-1 background_black">
             <p>
-                <i>Dernier message : </i>
+                <i class="text-white">Dernier message : </i>
                 <?= $service->afficherMessage($discussion["idDiscussion"]); ?>
             </p>
         </div>
@@ -163,7 +172,7 @@ foreach($discussions as $discussion)
         </div>
         <?php
 
-    /* Si l'utilisateur de la session n'a pas initié la discussion */
+    /* ------------ Si l'utilisateur de la session n'a pas initié la discussion ------------ */
     } else if ($discussion["idDestinataire"] == $idEmploye)
     {
         $idContact = $discussion["idEnvoyeur"];
@@ -171,39 +180,39 @@ foreach($discussions as $discussion)
         $contact = $objetUtilisateur->recupererUtilisateur($idContact);
 
         $dernierMessage = $objetMessage->recupererDernierMessage($discussion["idDiscussion"]);
-
         ?>
-
         <div class="container-fluid mt-100">
         <div class="row">
         <div class="col-md-12">
         <div class="card mb-4" style="border: none;">
         <div class="card-header text-white bg-dark">
-        <div class="media flex-wrap w-100 align-items-center">
-            <div class="rondAvatar">
-            <?php if(!empty($contact["avatar"])) { ?>
-                    <img src="<?=$contact["avatar"];?>" class="d-block ui-w-40 rounded-circle avatar">
-                <?php } else { ?>
-                    <img src="../images/avatar/avatarUtilisateur2.png" class="d-block ui-w-25 rounded-circle avatar">
-                    <?php } ?>
-            </div>
+            <div class="media flex-wrap w-100 align-items-center">
+                <div class="rondAvatar">
+                <?php if(!empty($contact["avatar"])) { ?>
+                        <img src="<?=$contact["avatar"];?>" class="d-block ui-w-40 rounded-circle avatar">
+                    <?php } else { ?>
+                        <img src="../images/avatar/avatarUtilisateur2.png" class="d-block ui-w-25 rounded-circle avatar">
+                        <?php } ?>
+                </div>
 
-            <div class="media-body ml-3">
-                Conversation avec :
-                <?php if($contact["idRole"]==2) { ?> <a style="color:blue;"> <?=$contact["prenom"] . " " . $contact["nom"];?></a>
-                <?php } if($contact["idRole"]==1){?><a><?=$contact["prenom"] . " " . $contact["nom"];?></a><?php } ?>
-                <div class="text-muted small">Dernière activité : <?=$service->dateFr($dernierMessage["max_date"]);?></div>
+                <div class="media-body ml-3">
+                    Conversation avec :
+                    <?php if($contact["idRole"]==2) { ?> <a style="color:blue;"> <?=$contact["prenom"] . " " . $contact["nom"];?></a>
+                    <?php } if($contact["idRole"]==1){?><a><?=$contact["prenom"] . " " . $contact["nom"];?></a><?php } ?>
+                    <div class="text-muted small">Dernière activité : <?=$service->dateFr($dernierMessage["max_date"]);?></div>
+                </div>
+                <!-- Top right -->
+                <a href="../traitements/supprimerDiscussion.php?idDiscussion=<?=$discussion["idDiscussion"];?>" class="icone_poubelle" onclick="return confirm('Êtes-vous sûr de vouloir supprimer la discussion ? Tous les messages seront également supprimés')">
+                    <i class="far fa-trash-alt"></i>
+                </a>
             </div>
-            <!-- Top right -->
-            <a href="../traitements/supprimerDiscussion.php" class="icone_poubelle">
-            <i class="far fa-trash-alt"></i>
-            </a>
         </div>
-        </div>
+        <!--
         <a href="discussion.php?id=<?=$discussion["idDiscussion"];?>" class="card_discussion">
-        <div class="card-body px-2 py-1">
+                -->
+        <div class="card-body px-2 py-1 background_black">
             <p>
-                <i>Dernier message : </i>
+                <i class="text-white">Dernier message : </i>
                 <?= $service->afficherMessage($discussion["idDiscussion"]); ?>
             </p>
         </div>
